requirements.txt
crewai
crewai-tools
duckduckgo-search
langchain-community
langchain
langchain-huggingface
langchain-chroma
sentence-transformers
python-dotenv
tavily-python
streamlit
mcp
pydantic-ai
